Refactor portfolio GUI rendering
